[Test] fix return type

+++ test/release/0-0-1.sol
@@ -17,7 +17,7 @@ contract SafeMath {
 
     function mul(uint256 a, uint256 b) public pure returns (uint256) {
         if (a == 0) {
-            return uint64(0);
+            return uint256(0);
         }
 
         uint256 c = a * b;
